Target user table and name column specifically
There is a config var called "user_name_column" that will query against
the parent application's "User" (or whatever) table and the column
containing the name they might be @ notified.

So whenever I do "hey @jim what's going on", thredded will peer into
`users.username`, or `users.name`, or `admins.first_name` ... what have
you ... and look for the user record to act on (link the name in the
post, email them, etc).

The issue here is the default in the config is too lax, and doesn't have
as clear an explanation.

There should *probably* be:

1. More clear documentation.
2. An exception raised upon runtime if that user table and column is not found.